placed functions into header file (main.h) and removed the main() function from _printf.c
