+++ main.h
@@ -1,0 +1,10 @@
+
+
+int isin(char *string, const char c);
+
+int count_valid_formatters(const char *string, ...);
+
+int printnum_as_chars(long n);
+
+int _printf(const char *string, ...);
+
